+++ NEWS.md
@@ -5,6 +5,7 @@
 - The `predictProbs()` and `predictChoices()` functions were depreciated.
 - A new `predict.logitr()` method was added for making probability and choice predictions from logitr class objects.
 - Added new `fitted.logitr()` and `residuals.logitr()` methods.
+- Added optional `predict` argument to the main `logitr()` function which controls whether predicted probabilities, fitted.values, and residuals are included in the returned object. Default setting is TRUE.
 - Changed the name of the coefficients vector in the returned object from "coef" to "coefficients" to be consistent with other packages.
 
 ## Bugs: